🔀 :: API v2로 모두 변경 ..

+++ Projects/Domain/BaseDomain/Interface/ReviewEntity.swift
@@ -11,13 +11,13 @@ import Foundation
 public struct ReviewEntity{
     let id: Int
     let content: String
-    let rating: Double
+    let rating: Int
     let images: [String]
     let statusCode: Int
     let message: String
     
     
-    public init(id: Int, content: String, rating: Double, images: [String], statusCode: Int, message: String) {
+    public init(id: Int, content: String, rating: Int, images: [String], statusCode: Int, message: String) {
         self.id = id
         self.content = content
         self.rating = rating
